+++ pac4j-core/src/main/java/org/pac4j/core/credentials/extractor/CredentialsExtractor.java
@@ -18,6 +18,8 @@ public interface CredentialsExtractor {
      * Extract the {@link Credentials} from a {@link org.pac4j.core.context.WebContext}
      * and return <code>Optional.empty()</code> if no credentials are present
      * or throw a {@link org.pac4j.core.exception.CredentialsException} if it cannot get it.
+     * Note that returning an empty {@code Optional} or throwing a {@code CredentialsException} will block
+     * any further authentication attempt. Any other exception will allow a retry.
      *
      * @param ctx the current context
      * @return the credentials (optional)